refactor(admin): extract EmailAnalyticsTab into its own file
